feat: implement Layers class for managing term layers and update references in compiler and grammar
